+++ examples/keyboard.rs
@@ -57,7 +57,7 @@ fn main() -> tetra::Result {
     let ctx = &mut ContextBuilder::new()
         .title("Keyboard Input")
         .size(160, 144)
-        .scale(4)
+        .window_scale(4)
         .resizable(true)
         .quit_on_escape(true)
         .build()?;